Map and Folder setup
Folders have been setup. Folders Created: "Materials": containing materials for structures, terrain etc, "Models" containing the Blender models used in the game, "Prefabs":  Containing the prefabs created in the unity editor. "Textures": containing the textures used in the game. "Resources": containing the prefabs that are going to be instantiated over the network.
The battlefield has been setup in the Battlefield Scene, and a MainMenu scene has been created for the Lobby browser.