Implemented simple PlayerController
Created PlayerController.cs,
PlayerController has Update Method, MovePlayer method to move the Player object. TryPlayerJump method casts a ray that checks if the Player object has contact with the ground and the Jump using a Impulse.
Updated CameraController so that the Player can now control the Camera when not in Spectator mode.