BattlUI Elements
BattleUI elements has been created in the Battlefield scene.
The BattleUI elements has been scripted so that they are updated when Player takes damage, when a Player fires his/hers weapon, when Player picks up MedPack, when Player picks up Ammo pack, and when Player Gets a Kill or when a Player wins the Battle.